+++ src/store/context.ts
@@ -1,0 +1,11 @@
+import type { ExtensionContext } from 'vscode'
+
+let ctx: ExtensionContext
+
+export function initExtensionContext(newCtx: ExtensionContext): void {
+  ctx = newCtx
+}
+
+export function getExtensionContext(): ExtensionContext {
+  return ctx
+}